requirements.txt
cachetools
fastapi
gitpython
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-cloud-bigquery
#google-cloud-billing
#google-cloud-essential-contacts==1.1.0
google-cloud-firestore
google-cloud-logging
hvac
pydantic
python-dotenv
requests
urlparse2
uvicorn[standard]
